Some Validations added to AccountController.java class

+++ OrangeBookApp/src/main/java/com/orange/spring/controller/AccountController.java
@@ -25,39 +25,76 @@ public class AccountController {
 	@Autowired
 	private AccountService accountService;
 	
-	  /*---get all accounts---*/
+	  /*--- (GET ALL) get all accounts---*/
 	   @GetMapping("/account")
 	   public ResponseEntity<List<Account>> list() {
 	      List<Account> accounts = accountService.list();
 	      return ResponseEntity.ok().body(accounts);
 	   }
 	   
-	  /*---Add new account---*/
+	  /*---(POST) Add new account---*/
 	   @PostMapping("/account")
 	   public ResponseEntity<?> save(@RequestBody Account account) {
-		  System.out.println("the json value of Account is :::::: "+account);
-	      long id = accountService.save(account);
+		  // New Validation By ID
+		  Account account2 = null;
+		  if (account.getId() != null) {
+			  account2 = accountService.get(account.getId());
+			  if (account2 != null) {
+				  //System.out.println("** ERROR, EXIST Account ::::: "+account2.toString());
+				  return ResponseEntity.ok().body("** ERROR ** EXIST Account. NO Account has been saved with ID:" + account.getId());
+			  }
+		  }
+		  // New Validation by IBAN
+		  if (account.getAccount_iban() == null || account.getAccount_iban().isEmpty()) {
+			  return ResponseEntity.ok().body("** ERROR ** Account IBAN IS NOT VALID :" + account.getAccount_iban());
+		  } else {
+			  account2 = accountService.getByIBAN(account.getAccount_iban());
+			  if (account2 != null) {
+				  //System.out.println("** ERROR, EXIST Account ::::: "+account2.toString());
+				  return ResponseEntity.ok().body("** ERROR ** EXIST Account IBAN:"+ account2.getAccount_iban() +". NO Account has been saved with ID:" + account.getId());
+			  }
+		  }
+		  // OJO Validation
+		  long id = accountService.save(account);
 	      return ResponseEntity.ok().body("New Account has been saved with ID:" + id);
 	  }
 	  
-	  /*---Get an account by id---*/
+	  /*---(GET 1)Get an account by id---*/
 	  @GetMapping("/account/{id}")
 	  public ResponseEntity<Account> get(@PathVariable("id") long id) {
 	     Account account = accountService.get(id);
 	     return ResponseEntity.ok().body(account);
 	  }
 
-	   /*---Update an account by id---*/
+	   /*---(PUT) Update an account by id---*/
 	   @PutMapping("/account/{id}")
 	   public ResponseEntity<?> update(@PathVariable("id") long id, @RequestBody Account account) {
-		   accountService.update(id, account);
+		  // New Validation By ID
+		  Account account2 = null;
+		  if (account.getId() != null) {
+			  account2 = accountService.get(account.getId());
+			  if (id != account2.getId()) {
+				  return ResponseEntity.ok().body("** ERROR ** Account and ID Missmatch. NO Account has been updated with ID:" + (id) +" / "+account.getId()); 
+			  }
+			  account2 = accountService.get(account.getId());
+			  if (account2 == null) {
+				  //System.out.println("** ERROR, EXIST Account ::::: "+account2.toString());
+				  return ResponseEntity.ok().body("** ERROR ** Account DO NOT EXIST. NO Account has been updated with ID:" + account.getId());
+			  }
+		  }
+		  accountService.update(id, account);
 	      return ResponseEntity.ok().body("Account has been updated successfully.");
 	   }
 	   
-	   /*---Delete an account id---*/
+	   /*---(DEL) Delete an account id---*/
 	   @DeleteMapping("/account/{id}")
 	   public ResponseEntity<?> delete(@PathVariable("id") long id) {
-		   accountService.delete(id);
+		  // New Validation By ID
+		   if (accountService.get(id) == null) {
+				  //System.out.println("** ERROR, EXIST Account ::::: "+account2.toString());
+				  return ResponseEntity.ok().body("** ERROR ** Account DO NOT EXISTT. NO Account has been deleted with ID:" + id);
+		  }
+		  accountService.delete(id);
 	      return ResponseEntity.ok().body("Account has been deleted successfully.");
 	   }
 }
